Move scabbard CommitHashStore to its own module
Move the `CommitHashStore` and related modules to a `commit_hash`
submodule in the store module. Re-export the necessary modules from
the store module for backwards compatibility.
