+++ services/scabbard/libscabbard/src/store/commit_hash/diesel/mod.rs
