+++ services/scabbard/libscabbard/src/store/commit_hash/diesel/schema.rs
