+++ services/scabbard/libscabbard/src/store/commit_hash/error.rs
